web_celery1/requirements.txt
asgiref==3.7.2
Django==4.2.7
pytz==2023.3.post1
sqlparse==0.4.4
celery==5.3
django_celery_beat
djangorestframework==3.14